feat: Separate ItemScreenshotFeature from item encoding, make ChatItemFeature have a separate keybind
