+++ common/src/main/resources/assets/wynntils/lang/en_us.json
@@ -121,6 +121,10 @@
   "feature.wynntils.changelog.name": "Changelog",
   "feature.wynntils.chatCoordinates.description": "Adds the ability to click on coordinates in chat to set your compass to that location.",
   "feature.wynntils.chatCoordinates.name": "Chat Coordinates",
+  "feature.wynntils.chatItem.chatItemError": "Failed to encode item!",
+  "feature.wynntils.chatItem.chatItemMessage": "Click here to copy the item for chat!",
+  "feature.wynntils.chatItem.chatItemTooltip": "Paste this text in chat to display your item to other Wynntils users",
+  "feature.wynntils.chatItem.chatItemUnidentifiedError": "Cannot make chat link of unidentified gear!",
   "feature.wynntils.chatItem.description": "Adds the ability to share items in chat.",
   "feature.wynntils.chatItem.name": "Chat Item",
   "feature.wynntils.chatMention.aliases.description": "Any aliases you wish to get mentioned for in addition to your username. Names should be separated with an single comma.",
@@ -643,9 +647,6 @@
   "feature.wynntils.itemLock.blockAllActionsOnLockedItems.name": "Block All Actions on Locked Items",
   "feature.wynntils.itemLock.description": "Adds the ability to lock items in your inventory to prevent dropping or interacting with them.",
   "feature.wynntils.itemLock.name": "Item Lock",
-  "feature.wynntils.itemScreenshot.chatItemError": "Cannot make chat link of unidentified gear!",
-  "feature.wynntils.itemScreenshot.chatItemMessage": "Click here to copy the item for chat!",
-  "feature.wynntils.itemScreenshot.chatItemTooltip": "Paste this text in chat to display your item to other Wynntils users",
   "feature.wynntils.itemScreenshot.copy.error": "Error while copying screenshot to clipboard",
   "feature.wynntils.itemScreenshot.copy.mac": "Item screenshot cannot be copied to clipboard on macOS.",
   "feature.wynntils.itemScreenshot.copy.mac.clickHere": "click here.",
