+++ common/src/main/resources/assets/wynntils/lang/zh_cn.json
@@ -467,9 +467,9 @@
   "feature.wynntils.itemLock.blockAllActionsOnLockedItems.name": "阻挡所有针对锁定物品的操作",
   "feature.wynntils.itemLock.description": "令你可以锁定你物品栏中的特定物品，防止其掉落以及交互。",
   "feature.wynntils.itemLock.name": "物品锁定",
-  "feature.wynntils.itemScreenshot.chatItemError": "无法为未鉴定装备生成聊天链接！",
   "feature.wynntils.itemScreenshot.chatItemMessage": "点击此处即可复制物品信息到聊天中！",
   "feature.wynntils.itemScreenshot.chatItemTooltip": "将这一文本粘贴至聊天中，即可向其他 Wynntils 用户展示你的物品。",
+  "feature.wynntils.itemScreenshot.chatItemUnidentifiedError": "无法为未鉴定装备生成聊天链接！",
   "feature.wynntils.itemScreenshot.copy.error": "将截图复制到剪贴板中时发生了错误",
   "feature.wynntils.itemScreenshot.copy.mac": "无法在 macOS 中将截图截图复制到剪贴板中。",
   "feature.wynntils.itemScreenshot.copy.mac.clickHere": "点击此处，",
